[19] Move mine func to internal directory
